Alterando codigo exemplo

+++ Testes com Java/Mockito/codigo/bookapi/src/main/java/com/exemplo/bookapi/repository/BookRepository.java
@@ -1,14 +1,15 @@
 package com.exemplo.bookapi.repository;
 
-import com.exemplo.bookapi.entities.Book;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
 import org.springframework.stereotype.Repository;
 
+import com.exemplo.bookapi.entity.Book;
+
 @Repository
-public interface BookRepository extends CrudRepository<Book, Long> {
+public interface BookRepository extends CrudRepository<Book, Integer> {
 
     Book findByIsbn(String isbn);
